+++ docs/content/en/changelog.md
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Unreleased
 
-- Upcoming feature will be reported here
+**Added**
+
+- `toHaveJson()` expectation
 
 ### [v1.3.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.2.0...v1.3.0) - 2021-15-10